Test auth for users
